requirements.txt
altair==5.5.0
fdb==2.0.3
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.46.0

